chore(ci): integration tests (#243)
* Revert "chore: remove integration tests (#196)"

This reverts commit bab7e2d672bf9459645c01201b37e6fcfd257521.

* add Makefile, prepare CI matrix dynamically

* fix set-matrix command

* add install libusb step in CI for integration tests

* clone testdata recursively

* update README.md and improve Makefile

* add missing depth and recursive git flags

* better Makefile formatting

* use u128 in fuzz/mean_gas

* use release build

* run e2e on master push only

+++ evm-adapters/src/fuzz.rs
@@ -169,7 +169,7 @@ impl FuzzedCases {
 
     /// Returns the average gas use of all test cases
     pub fn mean_gas(&self) -> u64 {
-        self.cases.iter().map(|c| c.gas).sum::<u64>() / self.cases.len() as u64
+        (self.cases.iter().map(|c| c.gas as u128).sum::<u128>() / self.cases.len() as u128) as u64
     }
 
     pub fn highest(&self) -> Option<&FuzzCase> {
